refactor: ProblemPane
