--- src/components/organisms_new/ErrorPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ErrorPane';
