+++ src/components/organisms_new/PaneManager/PaneManager.tsx
@@ -14,7 +14,7 @@ import {useMainPaneDimensions} from '@utils/hooks';
 import {ResizableColumnsPanel, ResizableRowsPanel} from '@monokle/components';
 import {activeProjectSelector} from '@shared/utils/selectors';
 
-import ErrorPane from '../ErrorPane';
+import ProblemPane from '../ProblemPane';
 import StartPage from '../StartPage';
 import * as S from './PaneManager.styled';
 import PaneManagerLeftMenu from './PaneManagerLeftMenu';
@@ -106,7 +106,7 @@ const NewPaneManager: React.FC = () => {
                     ) : currentActivity?.name === 'dashboard' ? (
                       <Dashboard />
                     ) : currentActivity?.name === 'validation' ? (
-                      <ErrorPane />
+                      <ProblemPane />
                     ) : (
                       <ActionsPane />
                     )
