+++ src/components/organisms_new/ProblemPane/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ProblemPane';
