+++ src/components/organisms_new/index.ts
@@ -1,10 +1,10 @@
 export {default as CompareSyncPane} from './CompareSyncPane';
 export {default as CreateProjectModal} from './CreateProjectModal';
-export {default as ErrorPane} from './ErrorPane';
 export {default as LearnPage} from './LearnPage';
 export {default as NavigatorPane} from './NavigatorPane';
 export {default as NewProject} from './NewProject';
 export {default as PaneManager} from './PaneManager';
+export {default as ProblemPane} from './ProblemPane';
 export {default as SettingsPane} from './SettingsPane';
 export {default as StartPage} from './StartPage';
 export {default as TemplateExplorer} from './TemplateExplorer';
